update analysis mysql rag

+++ ai/agents/agent_instance_util.py
@@ -10,7 +10,8 @@ from ai.agents.agentchat import (UserProxyAgent, GroupChat, AssistantAgent, Grou
                                  PythonProxyAgent, BIProxyAgent, TaskPlannerAgent, TaskSelectorAgent, CheckAgent,
                                  ChartPresenterAgent)
 from ai.backend.base_config import CONFIG
-# from ai.agents.agentchat.contrib import RetrieveAssistantAgent, RetrievePythonProxyAgent, RetrieveUserProxyAgent
+from ai.agents.agentchat.contrib import RetrieveAssistantAgent, RetrievePythonProxyAgent, RetrieveUserProxyAgent
+import os
 
 max_retry_times = CONFIG.max_retry_times
 language_chinese = CONFIG.language_chinese
@@ -216,8 +217,9 @@ class AgentInstanceUtil:
                 print(content)
             else:
                 print('self.rag_doc : ', self.rag_doc)
-                with open(self.rag_doc, 'w') as output_file:
-                    output_file.write(str(message))
+                if not os.path.exists(self.agent_instance_util.rag_doc):
+                    with open(self.rag_doc, 'w') as output_file:
+                        output_file.write(str(message))
 
                 self.is_rag = True
 
@@ -761,15 +763,15 @@ class AgentInstanceUtil:
 
     def get_agent_retrieve_database_describer(self):
         pass
-        # database_describer = RetrieveAssistantAgent(
-        #     name="retrieve_database_describer",
-        #     system_message="""data_describer.You are a data describer, describing in one sentence your understanding of the data selected by the user. For example, the data selected by the user includes X tables, and what data is in each table.""",
-        #     llm_config=self.gpt4_turbo_config,
-        #     user_name=self.user_name,
-        #     websocket=self.websocket,
-        #     openai_proxy=self.openai_proxy,
-        # )
-        # return database_describer
+        database_describer = RetrieveAssistantAgent(
+            name="retrieve_database_describer",
+            system_message="""data_describer.You are a data describer, describing in one sentence your understanding of the data selected by the user. For example, the data selected by the user includes X tables, and what data is in each table.""",
+            llm_config=self.gpt4_turbo_config,
+            user_name=self.user_name,
+            websocket=self.websocket,
+            openai_proxy=self.openai_proxy,
+        )
+        return database_describer
 
     def get_agent_bi_proxy(self):
         """ BI proxy """
@@ -809,20 +811,20 @@ class AgentInstanceUtil:
 
     def get_agent_retrieve_planner_user(self, is_log_out=True, report_file_name=None, docs_path=None):
         """Disposable conversation initiator, no reply"""
-        # retrieve_planner_user = RetrieveUserProxyAgent(
-        #     name="retrieve_planner_user",
-        #     max_consecutive_auto_reply=0,  # terminate without auto-reply
-        #     human_input_mode="NEVER",
-        #     websocket=self.websocket,
-        #     is_log_out=is_log_out,
-        #     openai_proxy=self.openai_proxy,
-        #     report_file_name=report_file_name,
-        #     retrieve_config={
-        #         "task": "qa",
-        #         "docs_path": docs_path,
-        #     },
-        # )
-        # return retrieve_planner_user
+        retrieve_planner_user = RetrieveUserProxyAgent(
+            name="retrieve_planner_user",
+            max_consecutive_auto_reply=0,  # terminate without auto-reply
+            human_input_mode="NEVER",
+            websocket=self.websocket,
+            is_log_out=is_log_out,
+            openai_proxy=self.openai_proxy,
+            report_file_name=report_file_name,
+            retrieve_config={
+                "task": "qa",
+                "docs_path": docs_path,
+            },
+        )
+        return retrieve_planner_user
 
     def get_agent_api_check(self):
         """Disposable conversation initiator, no reply"""
@@ -1001,6 +1003,22 @@ class AgentInstanceUtil:
 
     def get_agent_python_executor(self, report_file_name=None):
         python_executor = PythonProxyAgent(
+            name="python_executor",
+            system_message="python executor. Execute the python code and report the result.",
+            code_execution_config={"last_n_messages": 1, "work_dir": "paper"},
+            human_input_mode="NEVER",
+            websocket=self.websocket,
+            user_name=self.user_name,
+            default_auto_reply="TERMINATE",
+            # outgoing=self.outgoing,
+            # incoming=self.incoming,
+            db_id=self.db_id,
+            report_file_name=report_file_name,
+        )
+        return python_executor
+
+    def get_agent_retrieve_python_executor(self, report_file_name=None):
+        python_executor = RetrievePythonProxyAgent(
             name="python_executor",
             system_message="python executor. Execute the python code and report the result.",
             code_execution_config={"last_n_messages": 1, "work_dir": "paper"},
